Remove unnecesary "as" casts.
- "0 as *mut _" -> "ptr::null_mut()"; and the same goes for ptr::null()
- Simplify "0 as c_int" to "0" when the context already expects c_int
- And so on...

+++ src/librustc_back/dynamic_lib.rs
@@ -276,7 +276,7 @@ mod dl {
             None => {
                 let mut handle = ptr::null_mut();
                 let succeeded = unsafe {
-                    GetModuleHandleExW(0 as DWORD, ptr::null(), &mut handle)
+                    GetModuleHandleExW(0, ptr::null(), &mut handle)
                 };
                 if succeeded == 0 {
                     Err(io::Error::last_os_error().to_string())
