+++ src/libstd/io/cursor.rs
@@ -324,7 +324,7 @@ mod tests {
 
     #[test]
     fn test_buf_writer() {
-        let mut buf = [0 as u8; 9];
+        let mut buf = [0u8; 9];
         {
             let mut writer = Cursor::new(&mut buf[..]);
             assert_eq!(writer.position(), 0);
@@ -345,7 +345,7 @@ mod tests {
 
     #[test]
     fn test_buf_writer_seek() {
-        let mut buf = [0 as u8; 8];
+        let mut buf = [0u8; 8];
         {
             let mut writer = Cursor::new(&mut buf[..]);
             assert_eq!(writer.position(), 0);
@@ -374,7 +374,7 @@ mod tests {
 
     #[test]
     fn test_buf_writer_error() {
-        let mut buf = [0 as u8; 2];
+        let mut buf = [0u8; 2];
         let mut writer = Cursor::new(&mut buf[..]);
         assert_eq!(writer.write(&[0]).unwrap(), 1);
         assert_eq!(writer.write(&[0, 0]).unwrap(), 1);
